ShapePath: Replace implementation of toShapes().
