+++ docs/api/it/extras/core/ShapePath.html
@@ -75,16 +75,9 @@
 		<p>Collega una nuova [page:SplineCurve] al [page:ShapePath.currentPath currentPath].</p>
 
 
-		<h3>[method:Array toShapes]( [param:Boolean isCCW], [param:Boolean noHoles] )</h3>
+		<h3>[method:Array toShapes]()</h3>
 		<p>
-		isCCW -- Modifica la modalità di generazione degli hole e dei solidi<br/>
-		</p>
-		<p>
-      Converte l'array [page:ShapePath.subPaths subPaths] in un array di Shape. Per impostazione predefinita
-      le forme solide sono definite in senso orario (CW) e i fori sono definiti in senso antiorario (CCW).
-      Se isCCW è impostato su `true`, le forme sono capovolte.
-		<br/>
-
+      Converte l'array [page:ShapePath.subPaths subPaths] in un array di Shape.
 		</p>
