+++ docs/api/zh/extras/core/ShapePath.html
@@ -67,15 +67,9 @@
 			连接一个新的[page:SplineCurve]（样条曲线）到[page:ShapePath.currentPath currentPath]（当前路径）。</p>
 
 
-		<h3>[method:Array toShapes]( [param:Boolean isCCW] )</h3>
+		<h3>[method:Array toShapes]()</h3>
 		<p>
-		isCCW -- 更改实体形状和孔洞的生成方式。
-		</p>
-		<p>
-		将[page:ShapePath.subPaths subPaths]数组转换为到Shapes数组。默认情况下，实体形状按照顺时针（CW）来定义，孔洞按照逆时针（CCW）来定义。
-		如果isCCW被设置为true，则孔洞和实体形状的定义将被反转。
-		<br/>
-
+		将[page:ShapePath.subPaths subPaths]数组转换为到Shapes数组。
 		</p>
