+++ examples/webgl_geometry_extrude_shapes2.html
@@ -366,7 +366,7 @@
 						emissive: color
 					} );
 					const depth = depths[ i ];
-					const simpleShapes = path.toShapes( true );
+					const simpleShapes = path.toShapes();
 
 					for ( let j = 0; j < simpleShapes.length; j ++ ) {
