+++ examples/webgl_loader_svg.html
@@ -183,7 +183,7 @@
 								wireframe: guiData.fillShapesWireframe
 							} );
 
-							const shapes = SVGLoader.createShapes( path );
+							const shapes = path.toShapes();
 
 							for ( let j = 0; j < shapes.length; j ++ ) {
